+++ src/core/class/Product.js
